Move FormatManager and ResponseProcessor within the Statement namespace to indicate that they're internal/private.
